refactor: rewrite part 1
