+++ cognite/pygen/__init__.py
@@ -1,4 +1,4 @@
-from ._core.dms_to_python import SDKGenerator
+from ._core.generators import SDKGenerator
 from ._generator import generate_sdk, generate_sdk_notebook, write_sdk_to_disk
 from ._version import __version__
 from .utils.cdf import load_cognite_client_from_toml
