+++ cognite/pygen/utils/text.py
@@ -3,6 +3,30 @@ from __future__ import annotations
 import re
 
 import inflect
+from cognite.pygen.config import Naming, Case, Number
+
+
+def create_name(raw_name: str, naming: Naming) -> str:
+    """
+    Create a name from a raw name and following the given naming convention.
+    Args:
+        raw_name: The raw string to convert.
+        naming: The naming convention to follow.
+
+    Returns:
+        The converted name.
+
+    """
+    is_plural = naming.number == Number.plural
+    is_singular = naming.number == Number.singular
+    if naming == Case.pascal:
+        return to_pascal(raw_name, is_plural, is_singular)
+    elif naming == Case.snake:
+        return to_snake(raw_name, is_plural, is_singular)
+    elif naming == Case.camel:
+        return to_camel(raw_name, is_plural, is_singular)
+    else:
+        return raw_name
 
 
 def to_camel(string: str, pluralize: bool = False, singularize: bool = False) -> str:
