+++ tests/test_unit/generation/test_fields.py
@@ -1,6 +1,6 @@
 from cognite.client import data_modeling as dm
 
-from cognite.pygen._core.dms_to_python import APIGenerator
+from cognite.pygen._core.generators import APIGenerator
 
 
 def test_is_date_field(bid_view: dm.View) -> None:
